Python/requirements.txt
asttokens==2.0.5
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
cycler==0.10.0
executing==0.7.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
wincertstore==0.2
